adding more japanese options

+++ components/FlightList.js
@@ -57,12 +57,12 @@ export default function FlightList({ navigation }) {
     },
     rightInfoUpper: {
       flexDirection: "row",
-      justifyContent: "space-between"
+      justifyContent: "space-between",
     },
     tinyLogo: {
       width: 40,
       height: 40,
-      margin:3,
+      margin: 3,
     },
     deperatureArrival: {
       flexDirection: "row",
@@ -108,7 +108,7 @@ export default function FlightList({ navigation }) {
   const airlines = [];
   const years = [];
   const months = [];
-  const monNames = [
+  const enNames = [
     "January",
     "February",
     "March",
@@ -122,6 +122,28 @@ export default function FlightList({ navigation }) {
     "November",
     "December",
   ];
+  const jpNames = [
+    "1月",
+    "2月",
+    "3月",
+    "4月",
+    "5月",
+    "6月",
+    "7月",
+    "8月",
+    "9月",
+    "10月",
+    "11月",
+    "12月",
+  ];
+  let monNames = [];
+  if (state.language === "en") {
+    monNames = enNames;
+  }
+  if (state.language === "jp") {
+    monNames = jpNames;
+  }
+
   if (state.flightList.length > 0) {
     state.flightList.forEach((flight) => {
       let yr = flight.date.slice(0, 4);
@@ -141,8 +163,12 @@ export default function FlightList({ navigation }) {
 
   years.forEach((yr) => {
     if (filterItems.length === 0) {
+      let yLabel = "Year";
+      if (state.language === "jp") {
+        yLabel = "年";
+      }
       filterItems.push({
-        label: "Year",
+        label: yLabel,
         value: "yr",
         untouchable: true,
         textStyle: styles.labelHead,
@@ -155,8 +181,12 @@ export default function FlightList({ navigation }) {
     });
   });
 
+  let mLabel = "Month";
+  if (state.language === "jp") {
+    mLabel = "月";
+  }
   filterItems.push({
-    label: "Month",
+    label: mLabel,
     value: "mn",
     untouchable: true,
     textStyle: styles.labelHead,
@@ -170,8 +200,12 @@ export default function FlightList({ navigation }) {
     });
   });
 
+  let aLabel = "Airline";
+  if (state.language === "jp") {
+    aLabel = "航空会社";
+  }
   filterItems.push({
-    label: "Airline",
+    label: aLabel,
     value: "airline",
     untouchable: true,
     textStyle: styles.labelHead,
@@ -226,20 +260,41 @@ export default function FlightList({ navigation }) {
   };
 
   const confirmDelete = (id) => {
-    Alert.alert(
-      "Delete Flight",
-      "Are you sure you want to delete this flight?",
-      [
-        {
-          text: "Yes",
-          onPress: () => deleteFlight(id),
-        },
-        {
-          text: "No",
-        },
-      ]
-    );
+    if (state.language === "en") {
+      Alert.alert(
+        "Delete Flight",
+        "Are you sure you want to delete this flight?",
+        [
+          {
+            text: "Yes",
+            onPress: () => deleteFlight(id),
+          },
+          {
+            text: "No",
+          },
+        ]
+      );
+    }
+    if (state.language === "jp") {
+      Alert.alert(
+        "フライト削除",
+        "このフライトを削除してもよろしでしょうか？",
+        [
+          {
+            text: "はい",
+            onPress: () => deleteFlight(id),
+          },
+          {
+            text: "いいえ",
+          },
+        ]
+      );
+    }
   };
+  let placeholdertxt = "Filter by...";
+  if (state.language === "jp") {
+    placeholdertxt = "フィルター";
+  }
 
   function CreateList() {
     const arrayOfFlihtId = state.qrCodes.map((qrcode) => qrcode.flightID);
@@ -256,7 +311,12 @@ export default function FlightList({ navigation }) {
               activeOpacity={0.6}
             >
               <View style={styles.deleteBox}>
-                <Text style={styles.deleteText}>Delete</Text>
+                {state.language === "en" && (
+                  <Text style={styles.deleteText}>Delete</Text>
+                )}
+                {state.language === "jp" && (
+                  <Text style={styles.deleteText}>削除</Text>
+                )}
               </View>
             </TouchableOpacity>
           );
@@ -268,41 +328,42 @@ export default function FlightList({ navigation }) {
             dispatch({ type: "SetSelectedFlight", payload: l.id });
           }}
         >
-         
-            <View style={styles.flightLabel}>
-                <View style={styles.flightLabelUpper}>
-                    <View style={styles.logo}>
-                        <Image
-                        style={styles.tinyLogo}
-                        source={state.logo[l.airlineICAO]}
-                        ></Image>
-                    </View>
-                    <View style={styles.deperatureArrival}>
-                        <View >
-                            <Text style={styles.deperature}>{l.depAirport}</Text>
-                        </View>
-                        <View>
-                            <Text style={styles.tinyAirplane}> ✈︎ </Text>
-                        </View>
-                        <View>
-                            <Text style={styles.arrival}>{l.arrAirport}</Text>
-                        </View>
-                    </View>
-                    <View style={styles.rightInfo}>
-                        <View style={styles.rightInfoUpper}>
-                            <View>
-                                <Text>{l.flightNo}</Text>
-                            </View>
-                            <View>
-                                <Text>{arrayOfFlihtId.includes(l.id) ? <Text>✉️</Text> : <></>}</Text>
-                            </View>
-                        </View>
-                        <View>
-                            <Text>{moment(l.date).format("MMM Do YYYY")}</Text>
-                        </View>
-                    </View>
-                </View>   
+          <View style={styles.flightLabel}>
+            <View style={styles.flightLabelUpper}>
+              <View style={styles.logo}>
+                <Image
+                  style={styles.tinyLogo}
+                  source={state.logo[l.airlineICAO]}
+                ></Image>
+              </View>
+              <View style={styles.deperatureArrival}>
+                <View>
+                  <Text style={styles.deperature}>{l.depAirport}</Text>
+                </View>
+                <View>
+                  <Text style={styles.tinyAirplane}> ✈︎ </Text>
+                </View>
+                <View>
+                  <Text style={styles.arrival}>{l.arrAirport}</Text>
+                </View>
+              </View>
+              <View style={styles.rightInfo}>
+                <View style={styles.rightInfoUpper}>
+                  <View>
+                    <Text>{l.flightNo}</Text>
+                  </View>
+                  <View>
+                    <Text>
+                      {arrayOfFlihtId.includes(l.id) ? <Text>✉️</Text> : <></>}
+                    </Text>
+                  </View>
+                </View>
+                <View>
+                  <Text>{moment(l.date).format("MMM Do YYYY")}</Text>
+                </View>
+              </View>
             </View>
+          </View>
         </ListItem>
       </Swipeable>
     ));
@@ -333,14 +394,26 @@ export default function FlightList({ navigation }) {
             }}
             onPress={resetList}
           >
-            <Text
-              style={{
-                paddingLeft: 5,
-                paddingRight: 5,
-              }}
-            >
-              Clear Filter
-            </Text>
+            {state.language === "en" && (
+              <Text
+                style={{
+                  paddingLeft: 5,
+                  paddingRight: 5,
+                }}
+              >
+                Clear Filter
+              </Text>
+            )}
+            {state.language === "jp" && (
+              <Text
+                style={{
+                  paddingLeft: 5,
+                  paddingRight: 5,
+                }}
+              >
+                フィルター削除
+              </Text>
+            )}
           </TouchableOpacity>
           <DropDownPicker
             scrollViewProps={{
@@ -352,7 +425,7 @@ export default function FlightList({ navigation }) {
             min={0}
             max={10}
             defaultValue={""}
-            placeholder="Filter by..."
+            placeholder={placeholdertxt}
             containerStyle={{ width: 200 }}
             itemStyle={{
               justifyContent: "flex-start",
@@ -383,13 +456,20 @@ export default function FlightList({ navigation }) {
     return <QRScanner></QRScanner>;
   }
 
+  let hTitle = `${Auth.user.attributes.name}'s Flights`;
+  let h2 = "Flight Details";
+  if (state.language === "jp") {
+    hTitle = "フライトリスト";
+    h2 = "フライト情報";
+  }
+
   return (
     <Stack.Navigator initialRouteName="List">
       <Stack.Screen
         name="List"
         component={List}
         options={{
-          headerTitle: `${Auth.user.attributes.name}'s Flights`,
+          headerTitle: hTitle,
           headerStyle: {
             backgroundColor: "#298BD9",
           },
@@ -432,7 +512,7 @@ export default function FlightList({ navigation }) {
         name="Details"
         component={Details}
         options={{
-          headerTitle: "Flight Detail",
+          headerTitle: h2,
           headerTintColor: "#fff",
           headerStyle: {
             backgroundColor: "#298BD9",
